refact: update handlers
